+++ api/package.json
@@ -2,15 +2,17 @@
   "name": "portfolio-api",
   "version": "1.0.0",
   "description": "rachel reillys developer portfolio",
-  "main": "scr/index.js",
+  "main": "index.js",
   "scripts": {
-    "start": "node src/index.js"
+    "start": "node index.js"
   },
   "author": "rachel reilly",
   "license": "ISC",
   "dependencies": {
     "cors": "^2.8.5",
+    "dotenv": "^10.0.0",
     "express": "^4.17.1",
-    "helmet": "^4.6.0"
+    "helmet": "^4.6.0",
+    "node-fetch": "^2.6.1"
   }
 }
